pass HelloWorldController tests
